+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v1.1.0 [unreleased]
 
+- #576: Fix broken zoom on graphs that aren't the first.
 - #575: Add Varnish Layout
 - #574: Fix broken graphs on Postgres Layouts by adding aggregates.